Move chunk into its own header and src files
